refactor: decouple Z3 bridge support
