+++ docs/content/docs/research/11_tcb_audit.md
@@ -62,7 +62,7 @@ Every box's correctness obligation:
 | Checked/declaration boundary | `modules/verify/.../z3/ExpressionEncoder.scala` (`translateCheckedExpr`, `translateDeclarationExpr`) | **Trusted**, hand-written. Check bodies use the extracted translator only; declaration-level refinements may use the Scala fallback. |
 | `SmtTermBridge` | `modules/verify/.../z3/SmtTermBridge.scala` (`encodeFromSmtTerm`) | **Trusted**, hand-written. Converts extracted `smt_term` values to `Z3Expr` for the JNI driver. |
 | Declaration fallback | `modules/verify/.../z3/ExpressionEncoder.scala` (`translateDeclarationExprRaw`) | **Trusted but bypassed for sound checks**, invoked only for declaration-level expressions such as entity field constraints and type-alias `where` clauses. |
-| Sort inference, Z3 declaration management, set-of-T monomorphization, Skolem allocation, span threading | hand-written, split across `TranslationContext.scala`, `Declarations.scala`, `ExpressionEncoder.scala`, `RelationFrames.scala`, and `SmtTermBridge.scala` | **Trusted**, outside the soundness statement; runs before/around the bridge. |
+| Sort inference, Z3 declaration management, set-of-T monomorphization, Skolem allocation, span threading | hand-written, split across `TranslationContext.scala`, `Z3EncodingSupport.scala`, `Declarations.scala`, `ExpressionEncoder.scala`, `RelationFrames.scala`, and `SmtTermBridge.scala` | **Trusted**, outside the soundness statement; runs before/around the bridge. |
 | Z3 SMT solver (`com.microsoft.z3`) | external | **Trusted**, incomplete on quantifiers + nonlinear arithmetic + theory combinations. UNSAT verdicts are sound; SAT/Unknown can be solver artifacts. |
 | Counterexample decoder | `modules/verify/.../z3/CounterExample.scala` | **Trusted**, hand-written; outside soundness. |
 | Alloy backend | `modules/verify/src/main/scala/specrest/verify/alloy/*` | **Trusted**, separate bounded-model-checking story, rather than subject to `lower`. |
